fixed some issues

+++ upgrade/upgrade.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	COM "jpm/common"
 	"os"
+	"strings"
 )
 
 func Upgrade() {
@@ -15,7 +16,12 @@ func Upgrade() {
 		tempFile.Close()
 		COM.RunCMD(fmt.Sprintf(`curl -L -o "%s" https://cmd.jpmhub.org && start "upgrade" "C:\WINDOWS\system32\cmd.exe" /param="/c ""%s"""`, tempFile.Name(), tempFile.Name()), true)
 	} else {
-		// Simple background upgrade for Unix
+		// if /.sdkman/ is found in the exec path of this, propose to upgrade via sdkman
+		execDir, _ := os.Executable()
+		if strings.Contains(execDir, "/.sdkman/") {
+			println("\033[33mDetected sdkman installation of jpm. To upgrade, run:\n\n  sdk upgrade jpm\n\033[0m")
+			return
+		}
 		COM.RunScript(`curl -L -o s.sh https://sh.jpmhub.org && sh s.sh; rm -f s.sh`, true)
 	}
